Restructures renderers subpackage.
