+++ src/sherwood/events/animator.py
@@ -8,7 +8,7 @@ from types import TracebackType
 from typing import Any, Dict, Iterator, List, NamedTuple, Optional, Set, Tuple, Type
 
 from ..trees.base import Node, Tree
-from ..utils.draw import Renderer, draw_tree
+from ..renderers import Renderer, draw_tree
 from .base import AnimationNode, Bus, Event
